requirements.txt
# API e raccolta dati
requests>=2.28.1
websocket-client>=1.4.0
ccxt>=2.8.0
pycoingecko>=2.2.0
newsapi-python>=0.2.6
feedparser>=6.0.10

# Database
redis>=4.3.4
sqlalchemy>=1.4.40

# Analisi
pandas>=1.5.1
numpy>=1.23.4
scikit-learn>=1.1.3
nltk>=3.7

# Visualizzazione
plotly>=5.11.0
dash>=2.7.0
matplotlib>=3.6.2

# Sicurezza e logging
python-dotenv>=0.21.0
cryptography>=38.0.3
loguru>=0.6.0

# Test
pytest>=7.2.0
pytest-mock>=3.10.0

# Utilità
tqdm>=4.64.1
schedule>=1.1.0
PyJWT